requirements.txt
Flask==2.3.2
gunicorn==21.2.0
Flask-Caching==2.0.2
requests==2.31.0
pycryptodome==3.19.0
colorama==0.4.6
protobuf==4.24.4
